Swap the physical/logical mountings to fix EFS mount point
Swap the tasks for the physical and logical mountings so that the
logical is first, ensuring that the /mnt mount (i.e. storage/system as
the root) is mounted first and then the /efi filesystem can be mounted
inside that (rather than outside first).

This will fix issues with the creation of the UEFI stub kernels being
installed on the wrong filesystem.