Refactor libvirtd role to support creating a qemu partition
Refactor the libvirtd role to support creating an LVM volume for a qemu
partition as libvirt saves the virtual machines rather than stop them
and this consumes drive space, crashing libvirt on restart.

A 64GB partition is created which roughly equals the amount of RAM on a
host, allowing all running machines to be saved with space to spare.