Add roles for nginx and a caching proxy virtual host
Add roles for nginx, which will install and add the basic configuration
for nginx, with a set of standard and sane defaults, as well as a role
to configure the filesystem and a virtual host to manage caching through
a proxy the Arch Linux repositories and packages.

Also update the pacman configuration for bootstrapping and baselining to
use the new endpoints for the caching proxy.