Add a role for the configuration of Netdata for monitoring.
Add a role which will deploy and configure Netdata for the monitoring of
all servers in the Lab environment. This will set up three sets of
parent nodes (one for each environment) and then set up all currently
known child nodes to monitor their own data and then forward it to
parent node.

The parent nodes are set up with firewall rules for inbound traffic, and
local filesystem updates for storage. The nginx role is also updated to
support the configuration of monitoring against it by Netdata.