Refactor the Taskfile for dynamic Playbook handling
Refactor the Taskfile to support dynamic Ansible Playbook handling,
using the second component of the task to automatically look up the name
of the Playbook to run.

Additionally, create a test: set of tasks to run --check and --dry-run
against the Playbook for checking before running via the play: tasks.