Refactor linting for Ansible files
Refactor the linting for Ansible files by removing ansible-lint from the
critical path for most checks (it can normally take up to 60 seconds to
run, and many of the checks are now managed through jsonschema).
Instead, expand the supported checks through jsonschema and update the
task that needs to be run by play: and test: tasks.