Refactor the configuration of Caddy for Proxmox
Refactor the way the configuration of Caddy is handled for Proxmox,
first by moving the configuration in the caddy role to be just a base
configuration with no specific hosts set. Then, create a new role
specifically for Proxmox to add additional configuration to host the
Proxmox endpoints.

Also, extend the upgrade Ansible Play to enable the upgrading of Caddy
alongside all other packages as Caddy must be explicitly set to upgrade,
and it requires a custom process due to the additional plugins included.