Add an Ansible Role for Authentik
Add an Ansible Role for the installation and configuration of the
Authentik service. This service is a little unusual in that it is only
distributed in a container, so create systemd service files to run the
application via podman.

Authentik TLS is not handled by this role, as it must be manually
handled by the web interface it seems, and TLS is not configured between
the database and the service yet as that will require the CA configured
first.