Add an Ansible Role for cloudflared
Add an Ansible Role for cloudflared to support the running of Cloudflare
Tunnels from the Authentik nodes. This will allow Authentik to be
"publicly" hosted without using the public IP address of this network,
and provide additional checks and protections not offered locally.

Additionally, expand the number of type of entries that can be added to
the hosts file on a server to help direct traffic from Cloudflare to the
Caddy service on each server.