Add an Ansible Role for Valkey
Add an Ansible Role for Valkey to provide the in-memory database engine
needed (originally) for Authentik. However, as of the 2025.10.01
release, Authentik no longer supports Redis (and hence Valkey) for
operation, so although most of the role has been written, it has not
been fully tested in installation. It may be required for future
services, so the work can be kept.