Implement RFC2136 support in BIND with extra logging
Implement fully the RFC2136 updates in BIND, with changes to key types,
and their handling within the views and zones of the configuration file
(and how they relate to the various master and slave types). Also add
additional logging for general monitoring (at warn level or greater) and
add the remaining domains to the configuration.