Implement Lego to create ACME certificates for PostgreSQL
Implement Lego as a service (with support for standalone or Caddy-based
operations) to send requests to the internal Certificate Authority to
generate certificates matching the host's name and IP addresses.

The lego role will manage the creation and renewal of the certificates,
while the postgresql role will create the configuration for lego and set
up the trigger for checks and renewals.

Lego was chosen for support for RFC8738, which allows IP addresses to be
validated against by the ACME server, and it's simpler dependencies, and
recommended install process.