Migrate step-ca to using a replicated database
Tests of the failover process for step-ca showed that ACME clients
record a state that includes the unique IDs used by both the account and
the certificate. Attempts to renew a certificate on C2 that was
generated by C1 (or vice versa) when C1 is not operational fail.

A test to migrate to a replicated database shows that, as long as the
database is shared, C2 can renew certificates created by C1 and C2/C1
alternately.

This setup also uses the PostgreSQL replication connection to issue
write commands and route them to the current primary server by setting
the PostgreSQL hostname for the step-ca service to the local, secondary
database on C2. Additionally, this required a fix in the template for
the step-ca service configuration to support IPv6 addresses in the
database URI.