Add support for Proxmox Datacenter Manager hosts
Add support for the bootstrapping and configuration of Promox Datacenter
Manager hosts. These need many of the same resources as Proxmox VE
nodes, but with some small differences, such as not configuring IP
forwarding or thunderbolt ports. Also, the port used to host the
interface is different and there is a need to deal with different
hostnames.